Fix QuickJS node memoization leak and layout portability
- Migrated QuickJS node caching from strong Map to native weak mapping
  using LibDOM user data with heap-specific keys.
- Resolved PRIsizet macro spacing and concatenation issues in
  layout_flex.c, layout.c, and css_fetcher.c for better portability.
- Fixed a critical typo and added missing casts in image_cache.c
  macros (FMTCHR/FMTPCHR) to address compiler warnings.
- Hardened QuickJS runtime initialization (qjsky_init_runtime) to
  handle LibDOM string failures gracefully.
- Improved Duktape binding stability with defensive closed_down checks.

+++ content/handlers/html/css_fetcher.c
@@ -223,8 +223,8 @@ static void html_css_fetcher_poll(lwc_string *scheme)
 
 			if (c->aborted == false) {
 				snprintf(header, sizeof header,
-					"Content-Length: %"PRIsizet,
-					dom_string_byte_length(c->item->data));
+					"Content-Length: %" PRIsizet ,
+					(size_t) dom_string_byte_length(c->item->data));
 				msg.type = FETCH_HEADER;
 				msg.data.header_or_data.buf =
 						(const uint8_t *) header;
